+++ src/A1_csv_to_json.py
@@ -46,6 +46,10 @@ from src.constants import (
     TYPE_BOOL,
     TYPE_STR,
     TYPE_NONE,
+    EXTRA_CSV_PARAMETERS,
+    WARNING_TEXT,
+    REQUIRED_IN_CSV_ELEMENTS,
+    DEFAULT_VALUE,
 )
 from src.constants_json_strings import (
     LABEL,
@@ -66,12 +70,20 @@ from src.constants_json_strings import (
     SOC_MIN,
     STORAGE_CAPACITY,
     MAXIMUM_CAP,
+    RENEWABLE_ASSET_BOOL,
+    RENEWABLE_SHARE_DSO,
 )
 from src.constants_json_strings import UNIT, VALUE, ENERGY_STORAGE
 
 
 class MissingParameterError(ValueError):
     """Exception raised for missing parameters of a csv input file."""
+
+    pass
+
+
+class MissingParameterWarning(UserWarning):
+    """Exception raised for missing new parameters of a csv input file, which will be set to default."""
 
     pass
 
@@ -241,25 +253,7 @@ def create_json_from_csv(
             )
         )
 
-    # check wether parameter maximumCap is availavle
-    # TODO in next version: add maximumCap to hardcoded parameter list in constants.py
-    # TODO create this as a function, so that in future also new parameters can be added
-    list_of_new_parameters = {
-        MAXIMUM_CAP: "allows setting a maximum capacity for an asset that is being capacity optimized (Values: None/Float). ",
-        "renewableAsset": "allows defining a energyProduction asset as either renewable (True) or non-renewable (False) source. ",
-        "renewable_share": "allows defining the renewable share of the DSO supply (Values: Float). ",
-    }
-
-    for new_parameter in list_of_new_parameters:
-        if new_parameter in df.index:
-            parameters.append(new_parameter)
-        else:
-            # todo this message should only be displayed in case that the parameter is actually supposed to be applied to each of the files. For maxCap, this is also not valid, eg. it should not be added to economic_data
-            logging.warning(
-                f"You are not using the parameter {new_parameter} for asset group {filename}, which "
-                + list_of_new_parameters[new_parameter]
-                + "In the upcoming version of the MVS, this parameter will be required."
-            )
+    parameters = check_for_newly_added_parameters(filename, df, parameters)
 
     # check parameters
     missing_parameters = []
@@ -484,6 +478,69 @@ def create_json_from_csv(
         single_dict2.update({filename: single_dict})
         return single_dict2
     return
+
+
+def check_for_newly_added_parameters(
+    filename, df, required_parameters, official_extra_parameters=EXTRA_CSV_PARAMETERS
+):
+    """
+    Checks if there are new parameters that should be in the csvs.
+    Adds them to the required list of parameters.
+
+    Parameters
+    ----------
+    filename: str
+        Defines the name of a csv input file (without the extension)
+    df: :pandas:`pandas.DataFrame<frame>`
+        Data frame read from one of the input files
+    required_parameters: list
+        Defines the required parameters
+    official_extra_parameters: dict
+        dict specifing allowed extra parameters that should be in the Data frame
+
+    Returns
+    -------
+    Adds new parameter to list of parsed parameters or returns a warning if a new parameter is not defined in the csv. The parameter will be set to it's default value.
+    """
+    # Loop through official extra parameters (i.e. not yet added to the REQUIRED_CSV_PARAMETERS)
+    for extra_parameter in official_extra_parameters:
+        # Check whether the extra parameter should be contained in the csv file named `filename`
+        if (
+            filename
+            in official_extra_parameters[extra_parameter][REQUIRED_IN_CSV_ELEMENTS]
+        ):
+            # Check if the extra parameter is indeed included in the csv file named `filename`
+            if extra_parameter not in df.index:
+                # Add default values for each of the columns in the df
+                default_values = {}
+                for column in df:
+                    default_values.update(
+                        {
+                            column: official_extra_parameters[extra_parameter][
+                                DEFAULT_VALUE
+                            ]
+                        }
+                    )
+                default_values = pd.Series(data=default_values, name=extra_parameter)
+                df.append(default_values, ignore_index=False)
+
+                # Display warning message if the extra parameter was not present in the csv file.
+                warnings.warn(
+                    MissingParameterWarning(
+                        f"You are not using the parameter {extra_parameter} for asset group {filename}, which "
+                        + official_extra_parameters[extra_parameter][WARNING_TEXT]
+                        + ". "
+                        + f"This parameter is set to it's default value {official_extra_parameters[extra_parameter][DEFAULT_VALUE]}, which can influence the results."
+                        + "In the next release, this parameter will required."
+                    )
+                )
+
+            if extra_parameter not in required_parameters:
+                # Now that the new parameter is in the df
+                # (optional with default values being added) add the new parameter to parameter list
+                required_parameters.append(extra_parameter)
+
+    return required_parameters
 
 
 def conversion(value, asset_dict, row, param, asset, filename=""):