+++ CHANGELOG.md
@@ -31,6 +31,8 @@ Here is a template for new release sections
  - `utils.py` module in `src` to gather the functions `find_input_folders
  ` and `compare_input_parameters_with_reference` which can be used to find and validate input
   folders (#398)
+- Code and test for checking for new parameters in csv and raising warning message if not defined (`A1.check_for_newly_added_parameters`). This then also adds a default value to the new parameter  (#384)
+- Exception if an energyVector does not have internal generation or consumption from a DSO, and is only supplied by energy conversion from another sector: renewable share = 0. (#384)
 
 ### Changed
 - Use selenium to print the automatic project report, `python mvs_report.py -h` for help (#356)
@@ -38,6 +40,9 @@ Here is a template for new release sections
 - Change relative folder path to absolute in tests files (#396)
 - Replace all variables wacc, discount_factor and project_lifetime in the project (#383)
 - Improve styling of the pfd report (#369)
+- `LIST_OF_NEW_PARAMETERS` renamed `EXTRA_CSV_PARAMETERS` and moved from `A1` to `constants.py
+` (#384)
+- Order of parameters in tests/inputs, fixed missing parameters  (#384)
 
 ### Removed